メッセージ投稿時の非同期通信を実装

+++ app/assets/javascripts/message.js
@@ -1,24 +1,32 @@
 $(document).on('turbolinks:load', function(){
   function buildHTML(message){
-    var content = message.content ? `${ message.content }` : "";
-    var image = message.image ? `<img src= ${ message.image }>` : "";
-    var html = `<div class="content__box-image">
-                  <div class="content__box-images style="back-ground-image: ${ image };">
-                  <span>投稿者 ${ nickname }
-                  <a data-method="get" href="messages/${ id }/edit">編集</a>
-                  <a rel="nofollow" data-method="delete" href="/messages/${ id }">削除</a>
-                  <p>${ content }</p>
+    var text = message.text ? `${ message.text }` : "";
+    var image = message.image ? `${ message.image }` : "";
+    var html = `<div class="content__box-image" data-message-id=${message.id}>
+                  <div class="content__box-images" style="background-image: url(${ image });"></div>
+                  <span>投稿者 ${ message.nickname }</span>
+                  <div class="content__box-images__btns">
+                    <div class="content__box-images__btns--edit-btn" id="edit">
+                      <a data-method="get" href="messages/${ message.id }/edit">編集</a>
+                    </div>
+                    <div class="content__box-images__btns--delete-btn" id="delete">
+                      <a data-confirm="削除しますか？" rel="nofollow" data-method="delete" href="/messages/${ message.id }">削除</a>
+                    </div>
+                  </div>
+                  <div class="content__box-image--text">
+                    <p>${ text }</p>
+                  </div>
                   <div class="content__box-btn">
-                    <a data-method="get" href="/messags/${ id }">詳細</a>
+                    <a data-method="get" href="/messages/${ message.id }">詳細</a>
                   </div>
                 </div>`;
     return html;
   }
  
-  $('.form-btn').on('submit', function(e){
+  $('.new-message').on("submit", function(e){
     e.preventDefault();
     var formData = new FormData(this);
-    var url = $(this).attr('action')
+    var url = $(this).attr('action');
     $.ajax({
       type: "POST",
       url: url,
@@ -29,14 +37,13 @@ $(document).on('turbolinks:load', function(){
     })
     .done(function(data){
       var html = buildHTML(data);
-      $('.content__box').append(html);
-      $('.content__box').animate({scrollTop: $('.content__box')[0].scrollHeight}, 'swing');
-      // $('#new_message')[0].reset();
-      $('.form-btn').prop('disabled', false);
+      $('.content__box').prepend(html).hide().fadeIn(1000)
+      $('.new-message')[0].reset();
+      $('.form-btn-new').prop('disabled', false);
     })
     .fail(function(){
       alert('メッセージエラー');
-      $('.form-btn').prop('disabled', false);
+      $('.form-btn-new').prop('disabled', false);
     })
   })
 })
